+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# Head
+
+- Fixed: enabled and disabled debug in `watch` mode.
+
 # 1.0.0 - 2016-12-26
 
 - Chore: initial public release.